task 4: does alx have hair loss insurance? i really need to look into turkie's hair transplant surgery

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
 **/
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int numbits = 0, xor = 0;
+	unsigned long int numbits = 0, xor = 0;
 
 	for (xor = n ^ m; xor > 0; xor >>= 1)
 		if (xor & 1)
